изменен тип

+++ equ-api/src/main/java/equ/api/terminal/TerminalDocumentType.java
@@ -7,9 +7,9 @@ public class TerminalDocumentType implements Serializable {
     public String name;
     public String analytics1;
     public String analytics2;
-    public Integer flag;
+    public Long flag;
     
-    public TerminalDocumentType(String id, String name, String analytics1, String analytics2, Integer flag) {
+    public TerminalDocumentType(String id, String name, String analytics1, String analytics2, Long flag) {
         this.id = id;
         this.name = name;
         this.analytics1 = analytics1;
